417: Query the database for capacity percentage (#10)
* 417: Query the database for capacity percentage

Add integration tests too

* 417: Add sdr, sdr_conversions and paroms to the capacity calculations

* 417: Split unit and integration test tasks

Travis, on github should only execute unit tests until we have an integration
environment that we can run the DB dependent tests on.

* 417: Parse promise to extract display table

* 417: Update unit tests to handle promises

Now that the getCapacityTable returns a promise. The unit tests need to handle
this. This was achieved by using `sinon-bluebird` as used in APVS but required
downgrading the version of `sinon` used.

* 417: Ensure integration tests are cleaning up after themselves

Needed to use a user that has sufficent priviledges to delete records.
Hence using MIGRATION_APP user in the integration tests.

* 417: Fixing typo and structure after review

* 417: Fix standard linting errors