334: Add sub-navigation bar.
A navigation item is very similar to a breadcrumb item. Therefore refactor the
Breadcrumb object to a Link object and use for both breadcrumbs and sub-nav.

Import the sub nav sass from the ux prototype, refactoring slightly to avoid
duplication.

Extend the link generator in order to allow link generation from tree references
or name and id references. This allows us to reuse the link generation
functionality in the sub nav builder.

Build the sub-navigation at the same time as we're building the breadcrumbs and
the title/sub-title, i.e. in the get-capacity-view service. This will have to be
done on all requests and therefore should be extracted out to reduce
duplication.