334: Call getSubNav from the route instead of the service
The getSubNav call is going to be common across all of our screens. So it should
be as self contained as possible.

In stead of building up a list of navigation items in the service and then
setting the active flag on one, based on the req.path parameter, in the route.
The subNav array, including the active flag, is built in the get-sub-nav
service.

The service is then called only in the route, where we have access to the path.
This reduces the footprint of this common functionality and will make it easier
to bring new routes on board.