698: Add national level capacity view
The caseload-capacity screen should be visable at a national level. This will
display the aggregated capacity for every offender manager in the system. As
there is only one instance of the national level, no id field is required.
However a dummy id `0` is used in order to keep the routing consistent.

Therefore the national level capacity screen is accessed from the path:

`/nps/0/caseload-capacity`