Revert end to end test debugging changes
The end to end tests are currently failing on Jenkins because they require a
headless browser, such as PhantomJS.

The long term plan for this will be to use saucelabs to run on a series of
browsers. We are content to wait for this change to be played instead of
installing the drivers required to use a headless browser in Jenkins.