Create one knex instance per user instead of one per file.
Before this change the application created a new knex instance, and therefore a
new connection pool, each time it was reference from a file. That means we
couldn't control the size of the pool for the application.

The web application should only create one pool during normal operation. This is
done in the new `knex.js` file in the root of the project.

The integration tests will require a different pool whose connections use a user
with higher permissions. To achieve this another `knex.js` file was created in
the `test/` directory.