1212: Remove getContractedHours as it references redundant fields.
The application was giving regions with 0 contracted hours e.g. `Stakeholder
Engagement London` the default number of contracted hours for a PO or PSO. There
are some instances however, where the number of contracted hours should indeed
be 0. This is when migrated data has been brought over and there are regions
with no LDUs associated with them.