Remove conversion of date to string as it already gets parsed.
In the get-workload-report-view a start and end date are passed as strings.
These are then passed to the query.

The tests were passing date objects and therefore tests were failing until they
were converted to strings.

In the implementation however, the dates are converted to strings before they're
passed meaning the view was erroring out.