Call caseload views with noexpand hint.
Also, now the views take the total cases from the tiers table, instead of the
workload table - meaning the integration tests had to be updated.

Finally, now that all the caseload views are indexed, these indexes need to be
maintained when an insert occurs this means the timeout of requests need to be
extended.